Split up Box tests to avoid too many classes exception from styled-components
